modules/database-loader: don't fail if no zookeeper-dump is provided

+++ modules/contrail-database-loader.nix
@@ -67,7 +67,7 @@ in {
     systemd.services.zookeeper = {
       preStart = optionalString (cfg.zookeeperDumpPath != null) ''
         mkdir -p /var/lib/zookeeper/version-2/
-        cp ${cfg.zookeeperDumpPath}/* /var/lib/zookeeper/version-2/
+        [ "$(ls -A ${cfg.zookeeperDumpPath})" ] && cp ${cfg.zookeeperDumpPath}/* /var/lib/zookeeper/version-2/
         chown -R zookeeper:nogroup /var/lib/zookeeper/version-2/
       '';
     };
